fix: 購入復元ボタンを設定画面に戻す
